orbit_coordinator/requirements.txt
google-adk==1.27.2
a2a-sdk
python-dotenv
google-cloud-firestore
